refactor: Stream 올바른 사용,counting()사용에 따른 Long변경
순수함수로써 동작하도록 리팩토링

변경 전:
종언 함수인 forEach에서 외부 상태를 변경함

변경 후:
collect(groupingBy())를 사용하여 외부에 영향을 받지 않도록 함

+++ src/main/java/lotto/controller/Controller.java
@@ -58,7 +58,7 @@ public class Controller {
     }
 
     private void statusControl(WinningLotto winningLotto, List<Lotto> lottoList, LottoMoney lottoMoney) {
-        Map<LottoRank, Integer> rankResults =
+        Map<LottoRank, Long> rankResults =
                 lottoService.getRankResults(winningLotto, lottoList);
         Double revenueRate = lottoService.getRevenueRate(rankResults, lottoMoney);
         outputView.printWinningStatus(rankResults, revenueRate);
